refactor(cache): use LRUCache instead of own cache
BREAKING CHANGE:
- remove Source#onParsedFile callback
