+++ package.json
@@ -66,6 +66,7 @@
     "copc": "^0.0.6",
     "earcut": "^3.0.0",
     "js-priority-queue": "^0.1.5",
+    "lru-cache": "^11.0.1",
     "pbf": "^4.0.1",
     "shpjs": "^6.1.0",
     "threads": "^1.7.0"
