+++ test/unit/sources.js
@@ -1,0 +1,11 @@
+import './source/source';
+import './source/filesource';
+
+import './source/tmssource';
+import './source/wfssource';
+import './source/wmssource';
+import './source/wmtssource';
+
+import './source/orientedimagesource';
+
+import './source/C3DTilesSource';
